group protocol-related code together
