+++ src/blocks/notify.rs
@@ -13,7 +13,7 @@ use crate::config::SharedConfig;
 use crate::errors::*;
 use crate::formatting::value::Value;
 use crate::formatting::FormatTemplate;
-use crate::input::{I3BarEvent, MouseButton};
+use crate::protocol::i3bar_event::{I3BarEvent, MouseButton};
 use crate::scheduler::Task;
 use crate::util::pseudo_uuid;
 use crate::widgets::text::TextWidget;
