+++ src/blocks/nvidia_gpu.rs
@@ -9,7 +9,7 @@ use crate::config::SharedConfig;
 use crate::config::{LogicalDirection, Scrolling};
 use crate::de::deserialize_duration;
 use crate::errors::*;
-use crate::input::{I3BarEvent, MouseButton};
+use crate::protocol::i3bar_event::{I3BarEvent, MouseButton};
 use crate::scheduler::Task;
 use crate::util::pseudo_uuid;
 use crate::widgets::text::TextWidget;
