+++ src/blocks/speedtest.rs
@@ -12,7 +12,7 @@ use crate::de::deserialize_duration;
 use crate::errors::*;
 use crate::formatting::value::Value;
 use crate::formatting::{FormatTemplate, RenderedWidget};
-use crate::input::{I3BarEvent, MouseButton};
+use crate::protocol::i3bar_event::{I3BarEvent, MouseButton};
 use crate::scheduler::Task;
 use crate::widgets::text::TextWidget;
 use crate::widgets::{I3BarWidget, State};
