+++ src/blocks/watson.rs
@@ -9,7 +9,7 @@ use crate::config::SharedConfig;
 use crate::de::deserialize_duration;
 use crate::de::deserialize_local_timestamp;
 use crate::errors::*;
-use crate::input::I3BarEvent;
+use crate::protocol::i3bar_event::I3BarEvent;
 use crate::scheduler::Task;
 use crate::util::xdg_config_home;
 use crate::widgets::text::TextWidget;
